ir: pack IrInstruction to one cache line (96 -> 64 bytes)
The row's id always equaled its index in the function's instruction
array; ir_instruction_self_id recovers it by pointer difference for the
~20 consumers that asked the row. The seven operation/order enums store
as u8 with their COUNT sentinels intact (compile-time checked to fit),
and target/immediate counts store as u16 — the switch, computed-goto,
inline-assembly, and aggregate lowerings now fail their existing error
paths at 64K+ entries instead of silently truncating. operand_count
stays u32. 1.22M instances on the self-host unit drop ~39 MB of
touched bytes and one line per visit.

+++ src/buster/tests/compiler/ir/ir_test.c
@@ -70,10 +70,10 @@ BUSTER_GLOBAL_LOCAL IrValidationResult ir_test_canonical_f80_constant(Arena* are
                                                                                          .immediates = immediates,
                                                                                          .canonical_type = f80,
                                                                                          .targets = targets,
-                                                                                         .target_count = target_count,
+                                                                                         .target_count = (u16)target_count,
                                                                                          .result = value,
                                                                                          .opcode = IR_OPCODE_CONSTANT_FLOAT,
-                                                                                         .immediate_count = immediate_count,
+                                                                                         .immediate_count = (u16)immediate_count,
                                                                                          .next = IR_INSTRUCTION_ID_INVALID,
                                                                                      },
                                                                         (IrSourceRange){0})
